finish refactor of persona list/info
