+++ src/pages/persona/[persona].tsx
@@ -1,4 +1,5 @@
 import React from "react";
+import Head from "next/head";
 import { useRouter } from "next/router";
 import { getPersona } from "../../data";
 import { BasicInfo, ElementsTable, NameTags, Persona, ShadowInfo, StatsTable, SkillTable } from "../../persona";
@@ -14,16 +15,21 @@ const PersonaInfo = () => {
   const persona = getPersona(name);
   if (!name || !persona) return null;
   return (
-    <Persona persona={persona}>
-      <h1>
-        <NameTags persona={persona} />
-      </h1>
-      <BasicInfo />
-      {persona.shadow && <ShadowInfo />}
-      <StatsTable />
-      <ElementsTable />
-      <SkillTable />
-    </Persona>
+    <>
+      <Head>
+        <title>{persona.name} - rNav</title>
+      </Head>
+      <Persona persona={persona}>
+        <h1>
+          <NameTags persona={persona} />
+        </h1>
+        <BasicInfo />
+        {persona.shadow && <ShadowInfo />}
+        <StatsTable />
+        <ElementsTable />
+        <SkillTable />
+      </Persona>
+    </>
   );
-}
+};
 export default PersonaInfo;
