--- src/pages/persona/index.ts
@@ -1,2 +1,0 @@
-import PersonaTable from "../../personaList";
-export default PersonaTable;
