+++ src/personaList/index.ts
@@ -1,4 +1,1 @@
-export * from "./personaTableConfig";
 export * from "./personaList";
-import { PersonaList } from "./personaList";
-export default PersonaList;
